# Memoize output of FileStore.Stats