###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass

This should significantly reduce the time spent planning compactions when there are large numbers of TSM files within a shard, by memoizing the output of `FileStore.Stats`. When new TSM files are added or removed the output of `FileStore.Stats` will need to be recalculated.


Before:

```
⇒  gt -run NONE -bench BenchmarkFileStore_Stats
BenchmarkFileStore_Stats-4   	    5000	    215589 ns/op	   82094 B/op	       2 allocs/op
PASS
ok  	github.com/influxdata/influxdb/tsdb/engine/tsm1	9.185s
```

After:

```
⇒  gt -run NONE -bench BenchmarkFileStore_Stats
BenchmarkFileStore_Stats-4   	20000000	        99.2 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/influxdata/influxdb/tsdb/engine/tsm1	15.388s
```

@jwilder @benbjohnson 

/cc @beckettsean