@jwilder 2016-10-20
The rename in your test fails on windows/appveyor because the file is open. Probably need to do something like: https://github.com/influxdata/influxdb/blob/master/tsdb/engine/tsm1/compact_test.go#L2175